requirements.txt
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
removebg==0.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
